Continued v1.1.0 Development

+++ Implementation/ReadySetResource/ReadySetResource/Areas/Apps/Controllers/HolidaysController.cs
@@ -13,22 +13,20 @@ using ReadySetResource.Models;
 using ReadySetResource.ViewModels;
 using Microsoft.AspNet.Identity;
 using Microsoft.AspNet.Identity.Owin;
+using System.Web.Routing;
 #endregion
 
 namespace ReadySetResource.Areas.Apps.Controllers
 {
-    /// <summary>
-    /// This is the holiday for the calendar application so that users can see their shift schedule.
-    /// </summary>
-    /// <seealso cref="System.Web.Mvc.Controller" />
+
+
+
+    [Authorize]
     public class HolidaysController : Controller
     {
-        #region Context
+        #region Context and Initialisers
         private ApplicationDbContext _context;
 
-        /// <summary>
-        /// Initializes a new instance of the <see cref="HolidaysController"/> class.
-        /// </summary>
         public HolidaysController()
         {
             _context = new ApplicationDbContext();
@@ -43,6 +41,21 @@ namespace ReadySetResource.Areas.Apps.Controllers
         [Authorize]
         public ActionResult Index(DateTime? week)
         {
+
+            #region Authorise App
+            var userId = User.Identity.GetUserId();
+            var user = _context.Users.SingleOrDefault(c => c.Id == userId);
+            var userType = _context.BusinessUserTypes.FirstOrDefault(t => t.Id == user.BusinessUserTypeId);
+            var appName = this.ControllerContext.RouteData.Values["controller"].ToString();
+            var app = _context.Apps.FirstOrDefault(a => a.Link == appName);
+            var accessType = _context.TypeAppAccesses.Where(t => t.AppId == app.Id).Where(t => t.BusinessUserTypeId == userType.Id).ToList();
+
+            if (accessType.Count == 0)
+            {
+                return RedirectToAction("NotAuthorised", "Account", new { area = "" });
+            }
+            #endregion
+
             DateTime weekBeginDate;
             if (week != null) { weekBeginDate = week.Value; }
             else { weekBeginDate = DateTime.Now.Date; }
@@ -61,6 +74,20 @@ namespace ReadySetResource.Areas.Apps.Controllers
         [Authorize]
         public ActionResult Add(DateTime? date)
         {
+            #region Authorise App
+            var userId = User.Identity.GetUserId();
+            var user = _context.Users.SingleOrDefault(c => c.Id == userId);
+            var userType = _context.BusinessUserTypes.FirstOrDefault(t => t.Id == user.BusinessUserTypeId);
+            var appName = this.ControllerContext.RouteData.Values["controller"].ToString();
+            var app = _context.Apps.FirstOrDefault(a => a.Link == appName);
+            var accessType = _context.TypeAppAccesses.Where(t => t.AppId == app.Id).Where(t => t.BusinessUserTypeId == userType.Id).ToList();
+
+            if (accessType.Count == 0)
+            {
+                return RedirectToAction("NotAuthorised", "Account", new { area = "" });
+            }
+            #endregion
+
             DateTime holidayDate;
             if (date != null) { holidayDate = date.Value; }
             else { holidayDate = DateTime.Now.Date; }
@@ -109,17 +136,23 @@ namespace ReadySetResource.Areas.Apps.Controllers
 
 
         #region Edit (View)
-        // GET: Dashboard/Edit
-        /// <summary>
-        /// Edits the specified holiday.
-        /// </summary>
-        /// <param name="holiday">The holiday.</param>
-        /// <returns>The view with the holidayVM</returns>
         [HttpGet]
         [Authorize]
         public ActionResult Edit(int holiday)
         {
+            #region Authorise App
+            var userId = User.Identity.GetUserId();
+            var user = _context.Users.SingleOrDefault(c => c.Id == userId);
+            var userType = _context.BusinessUserTypes.FirstOrDefault(t => t.Id == user.BusinessUserTypeId);
+            var appName = this.ControllerContext.RouteData.Values["controller"].ToString();
+            var app = _context.Apps.FirstOrDefault(a => a.Link == appName);
+            var accessType = _context.TypeAppAccesses.Where(t => t.AppId == app.Id).Where(t => t.BusinessUserTypeId == userType.Id).ToList();
 
+            if (accessType.Count == 0)
+            {
+                return RedirectToAction("NotAuthorised", "Account", new { area = "" });
+            }
+            #endregion
 
             //1 - Get BusinessUserType from current user and sets current user as .cshtml needs to check for business user type
             var currUserId = User.Identity.GetUserId();
@@ -141,7 +174,7 @@ namespace ReadySetResource.Areas.Apps.Controllers
                 StartHour = ActualHoliday.StartDateTime.Hour.ToString(),
                 StartMinute = ActualHoliday.StartDateTime.Minute.ToString(),
 
-                
+
 
                 UserId = ActualHoliday.UserId,
 
@@ -175,28 +208,11 @@ namespace ReadySetResource.Areas.Apps.Controllers
 
 
 
-        #region Test
-        [HttpGet]
-        [Authorize]
-        public ActionResult Test()
-        { 
-            return View();
-        }
-        #endregion
-
-
-
-
-
         //Methods for Holidays
         #region Populate Holidays Method
-        /// <summary>
-        /// Populates the holidays.
-        /// </summary>
-        /// <param name="weekBeginDate">The week begin date.</param>
-        /// <returns>The holidays view model</returns>
         private HolidaysViewModel PopulateHolidays(DateTime weekBeginDate)
         {
+
             //1 - Get BusinessUserType from current user and sets current user as holidays.cshtml needs to check for business user type
             var currUserId = User.Identity.GetUserId();
             var currBusinessUser = _context.Users.SingleOrDefault(c => c.Id == currUserId);
@@ -263,8 +279,23 @@ namespace ReadySetResource.Areas.Apps.Controllers
         [Authorize]
         public ActionResult AddHoliday(HolidayViewModel holidayVM)
         {
-            ApplicationUser user = new ApplicationUser();
-            if(holidayVM.UserId == null)
+
+            #region Authorise App
+            var userId = User.Identity.GetUserId();
+            var user = _context.Users.SingleOrDefault(c => c.Id == userId);
+            var userType = _context.BusinessUserTypes.FirstOrDefault(t => t.Id == user.BusinessUserTypeId);
+            var appName = this.ControllerContext.RouteData.Values["controller"].ToString();
+            var app = _context.Apps.FirstOrDefault(a => a.Link == appName);
+            var accessType = _context.TypeAppAccesses.Where(t => t.AppId == app.Id).Where(t => t.BusinessUserTypeId == userType.Id).ToList();
+
+            if (accessType.Count == 0)
+            {
+                return RedirectToAction("NotAuthorised", "Account", new { area = "" });
+            }
+            #endregion
+
+            user = new ApplicationUser();
+            if (holidayVM.UserId == null)
             {
                 holidayVM.UserId = User.Identity.GetUserId();
             }
@@ -313,7 +344,7 @@ namespace ReadySetResource.Areas.Apps.Controllers
                 }
             }
 
-            
+
 
 
             var currUserId = User.Identity.GetUserId();
@@ -381,8 +412,21 @@ namespace ReadySetResource.Areas.Apps.Controllers
         public ActionResult EditHoliday(HolidayViewModel holidayVM)
         {
 
+            #region Authorise App
+            var userId = User.Identity.GetUserId();
+            var user = _context.Users.SingleOrDefault(c => c.Id == userId);
+            var userType = _context.BusinessUserTypes.FirstOrDefault(t => t.Id == user.BusinessUserTypeId);
+            var appName = this.ControllerContext.RouteData.Values["controller"].ToString();
+            var app = _context.Apps.FirstOrDefault(a => a.Link == appName);
+            var accessType = _context.TypeAppAccesses.Where(t => t.AppId == app.Id).Where(t => t.BusinessUserTypeId == userType.Id).ToList();
+
+            if (accessType.Count == 0)
+            {
+                return RedirectToAction("NotAuthorised", "Account", new { area = "" });
+            }
+            #endregion
+
             //Finds the user and sets the holiday user and its id
-            ApplicationUser user = new ApplicationUser();
             user = _context.Users.SingleOrDefault(u => u.Id == holidayVM.UserId);
             Shift holiday = new Shift();
             holiday.User = user;
@@ -410,7 +454,7 @@ namespace ReadySetResource.Areas.Apps.Controllers
             holiday.EndDateTime = holiday.EndDateTime.AddDays(holidayVM.EndDate.Day - 1);
             holiday.EndDateTime = holiday.EndDateTime.AddMonths(holidayVM.EndDate.Month - 1);
             holiday.EndDateTime = holiday.EndDateTime.AddYears(holidayVM.EndDate.Year - 1);
-            
+
 
 
             //Checks to see if there are any changes to the holiday 
@@ -482,7 +526,7 @@ namespace ReadySetResource.Areas.Apps.Controllers
                 return RedirectToAction("Index", "Holidays", new { week = holidayVM.StartDate.Date });
             }
 
-            
+
 
 
         }
@@ -491,15 +535,23 @@ namespace ReadySetResource.Areas.Apps.Controllers
 
 
         #region AcceptHoliday
-        // POST: Calendar/DeleteHoliday
-        /// <summary>
-        /// Accepts the holiday.
-        /// </summary>
-        /// <param name="holiday">The holiday.</param>
-        /// <returns>The index action with the updated holidays list</returns>
         [Authorize]
         public ActionResult AcceptHoliday(int holiday)
         {
+            #region Authorise App
+            var userId = User.Identity.GetUserId();
+            var user = _context.Users.SingleOrDefault(c => c.Id == userId);
+            var userType = _context.BusinessUserTypes.FirstOrDefault(t => t.Id == user.BusinessUserTypeId);
+            var appName = this.ControllerContext.RouteData.Values["controller"].ToString();
+            var app = _context.Apps.FirstOrDefault(a => a.Link == appName);
+            var accessType = _context.TypeAppAccesses.Where(t => t.AppId == app.Id).Where(t => t.BusinessUserTypeId == userType.Id).ToList();
+
+            if (accessType.Count == 0)
+            {
+                return RedirectToAction("NotAuthorised", "Account", new { area = "" });
+            }
+            #endregion
+
             Holiday actualHoliday = _context.Holidays.SingleOrDefault(s => s.Id == holiday);
             actualHoliday.Accepted = "Accepted";
             _context.SaveChanges();
@@ -511,15 +563,23 @@ namespace ReadySetResource.Areas.Apps.Controllers
 
 
         #region DeclineHoliday
-        // POST: Calendar/DeleteHoliday
-        /// <summary>
-        /// Declines the holiday.
-        /// </summary>
-        /// <param name="holiday">The holiday.</param>
-        /// <returns>The index action with the updated holidays list</returns>
         [Authorize]
         public ActionResult DeclineHoliday(int holiday)
         {
+            #region Authorise App
+            var userId = User.Identity.GetUserId();
+            var user = _context.Users.SingleOrDefault(c => c.Id == userId);
+            var userType = _context.BusinessUserTypes.FirstOrDefault(t => t.Id == user.BusinessUserTypeId);
+            var appName = this.ControllerContext.RouteData.Values["controller"].ToString();
+            var app = _context.Apps.FirstOrDefault(a => a.Link == appName);
+            var accessType = _context.TypeAppAccesses.Where(t => t.AppId == app.Id).Where(t => t.BusinessUserTypeId == userType.Id).ToList();
+
+            if (accessType.Count == 0)
+            {
+                return RedirectToAction("NotAuthorised", "Account", new { area = "" });
+            }
+            #endregion
+
             Holiday actualHoliday = _context.Holidays.SingleOrDefault(s => s.Id == holiday);
             actualHoliday.Accepted = "Declined";
             _context.SaveChanges();
@@ -531,15 +591,25 @@ namespace ReadySetResource.Areas.Apps.Controllers
 
 
         #region DeleteHoliday
-        // POST: Calendar/DeleteHoliday
-        /// <summary>
-        /// Deletes the holiday.
-        /// </summary>
-        /// <param name="holiday">The holiday.</param>
-        /// <returns>The index action with the updated holidays list</returns>
+
         [Authorize]
         public ActionResult DeleteHoliday(int holiday)
         {
+
+            #region Authorise App
+            var userId = User.Identity.GetUserId();
+            var user = _context.Users.SingleOrDefault(c => c.Id == userId);
+            var userType = _context.BusinessUserTypes.FirstOrDefault(t => t.Id == user.BusinessUserTypeId);
+            var appName = this.ControllerContext.RouteData.Values["controller"].ToString();
+            var app = _context.Apps.FirstOrDefault(a => a.Link == appName);
+            var accessType = _context.TypeAppAccesses.Where(t => t.AppId == app.Id).Where(t => t.BusinessUserTypeId == userType.Id).ToList();
+
+            if (accessType.Count == 0)
+            {
+                return RedirectToAction("NotAuthorised", "Account", new { area = "" });
+            }
+            #endregion
+
             Holiday actualHoliday = _context.Holidays.SingleOrDefault(s => s.Id == holiday);
             _context.Holidays.Remove(actualHoliday);
             _context.SaveChanges();
@@ -547,5 +617,6 @@ namespace ReadySetResource.Areas.Apps.Controllers
             return RedirectToAction("Index", "Holidays", new { week = actualHoliday.StartDateTime });
         }
         #endregion
+
     }
 }
